# pyroscope.scrape: remove relabeling, remove discovered targets field, remove public labels field, use labels builder