--- a/internal/component/pyroscope/scrape/target_test.go
+++ b/internal/component/pyroscope/scrape/target_test.go
@@ -161,7 +161,6 @@ func Test_NewTarget_godeltaprof(t *testing.T) {
 	}), url.Values{})
 
 	require.NotEqual(t, withGodeltaprof.allLabels, withoutGodeltaprof.allLabels)
-	require.Equal(t, withGodeltaprof.publicLabels, withoutGodeltaprof.publicLabels)
 	assert.Equal(t, pprofMemory, withGodeltaprof.allLabels.Get(model.MetricNameLabel))
 	assert.Equal(t, pprofMemory, withoutGodeltaprof.allLabels.Get(model.MetricNameLabel))
 	assert.Equal(t, "/debug/pprof/heap", withoutGodeltaprof.allLabels.Get(ProfilePath))