<!--

CONTRIBUTORS GUIDE: https://github.com/grafana/alloy/blob/main/docs/developer/contributing.md#updating-the-changelog

If this is your first PR or you have not contributed in a while, we recommend
taking the time to review the guide. It gives helpful instructions for
contributors around things like how to update the changelog.

-->

#### PR Description
- remove relabeling logic, which was impossible to configure and therefore it was dead code.
- remove dropped targets logic as a result
- remove `Target.discoveredLabels` field as it was only used as debug information (was replaced with `Target.allLabels` in debug information). As a result, remove the second argument to `NewTarget` simplifying testing code.
- use `labels.Builder` for `populateLabels` to reduce some allocations
- use `filepath.Join` instead of `URL.joinPath` to save some allocations
- add a small benchmark
- remove publicLabels field. 
- calculate public labels hash with less allocations

~Found a bug: Updating labels, does not work. The updating logic was updating `DiscoveredTargets` which are not used when data is forwarded to `pyroscope.write`. Created an issue: #2628~ EDIT: never-mind, I think we should not update anything at all, unless we decide to change __meta labels dropping behavior, see #2629
```
Before
BenchmarkPopulateLabels-12    	   11810	    110312 ns/op	   77465 B/op	     661 allocs/op
BenchmarkPopulateLabels-12    	   10000	    106040 ns/op	   77466 B/op	     661 allocs/op
After
BenchmarkPopulateLabels-12    	   24166	     52106 ns/op	   30082 B/op	     280 allocs/op
BenchmarkPopulateLabels-12    	   23424	     54803 ns/op	   30082 B/op	     280 allocs/op
```

Note: allocations and performance are not a big issue here. The goal of the PR is to remove dead code.

#### Which issue(s) this PR fixes

<!-- Uncomment the following line if you want that GitHub issue gets automatically closed after merging the PR -->
<!-- Fixes #issue_id -->

#### Notes to the Reviewer

#### PR Checklist

<!-- Remove items that do not apply. For completed items, change [ ] to [x]. -->

- [x] CHANGELOG.md updated
- [ ] Documentation added
- [x] Tests updated
- [ ] Config converters updated